Release 0.9r0
OpenLB Release 0.9: Open Source Lattice Boltzmann Code
Krause, M.J. and Henn, T. and Mink, A. and Trunk, R. and Weisbrod, P. and Nathen, P. and Klemens, F. and Maier, M.-L.
10.5281/zenodo.3625941

+++ src/core/dataFields2D.cpp
@@ -26,15 +26,15 @@
  */
 #include "dataFields2D.h"
 #include "dataFields2D.hh"
-#include "latticeDescriptors.h"
-#include "latticeDescriptors.hh"
+#include "dynamics/latticeDescriptors.h"
+#include "dynamics/latticeDescriptors.hh"
 
 namespace olb {
 
 using namespace descriptors;
 
 template class ScalarField2D<double>;
-template class ScalarField2D<unsigned short>;
+template class ScalarField2D<int>;
 template class SequentialScalarFieldSerializer2D<double>;
 template class SequentialScalarFieldUnSerializer2D<double>;
